requirements.txt
pytest==7.4.4
requests==2.31.0
pytest-html==4.1.1
jsonschema==4.24.0
python-dotenv==1.1.0